images/daily-pipeline/upload_to_influxdb/requirements.txt
pandas
google-cloud-storage
influxdb-client
